Showing some text and playing with the data model
I've decided to start with a simple idea of having a model of messages in the app state.
The model is a vector of maps. Each map contains things like title, text, class, etc.
I'm not yet sure is this works or not but for now I like the simplicity of the idea.

I've added some boilerplate code to read the data of the message (which is now hardcoded in the init step) and show it on the screen.