Message that looks like a message
In order to have a more pleasant experience live-coding with figwheel I've styled the message UI component a bit so it would look like a proper message not just text floating in the void.

I've also changed message model's field text to be description. Text is to vague and will clash with css attributes/classes, which will not break anything but will make it harder to read.

P.S. Please don't laugh at my CSS :)