Close button added and works
Well, it's not really a button, it's just an X symbol that's causes the cursor to change on hover to look like a pointer so you know you can click it :)
I have also added a close-message event and a simple handler. It's just data work, so after fiddling with update-in and remove functions in the repl I've created a small function that does what I need.

Now you have two things to click on! :)