Using reg-event-db correctly
Still reading through re-frame docs to get a better understanding. It seems that reg-event-db will only pass db as the first param so I don't need to do destructuring like in reg-event-fx.

Also noticed that I had a part of the code left from the repl-fiddling. So the function was updating the local var instead of the app-state. Fixed. Still works.