Naive implementation of stackable messages
Currently the offset for the next message is calculated based on the current index (count) of the message and a hardcoded value of 75 (looks fine on my screen :)).

Next step: I need to change this to take into the account the height of the previous message instead of hardcoded 75 value.
Or maybe there's a magic CSS thing to take relative position from the previous element.
Or! Maybe I can use something even simpler, like padding-top! I like this idea, no fancy calculations needed. I'll try it next...