Stacking messages based on their height
OK, don't know why I thought I can do this with padding, probably was getting tired.
Decided to base the position of the new message on the position and the height of the previous one. This is needed because I want the height of the messages to be flexible (don't want to see scrollbars). This means we need to calculate the offset properly. Not a big deal, the calculation is simple. I'm using goog.dom to grab element's attributes so I've added id html attribute to each message component.

At some point I need to add some unit tests for the util namespace. That's the only thing that has any logic atm.