Fixed a stackable bug and added more message flavors
My previous calculation to stack messages on top of another had a bug: when you dismiss some messages the JS is broken
because it was looking for ID that didn't exist anymore. I was erroneously using message's index (position in messages vector) as a base for id instead of keep using index in array of message elements on the page. Changed that by grabbing all elements of class "message" from the html (nodes) and searching in that array by the same index that I've used in the app-state.

Also added two more flavors: Warning Message and Error Message. Added playground buttons to generate them to test and enjoy the tehcnicolor rainbow!