Self-destruction implemented based on setInterval and global state
The self-destruction system works like this:
- when a message is created :ttl value can be specified in ms
- the ttl value is used to populate :destroy-at attribute of the related message
- every second timer event is dispatched with the value of current time
- the timer event handler runs through messages removing everything that has :destroy-at value older than the current time

In order to play-test this I've updated the playground with additional text field that you can use to specify the TTL in ms for each message.

I've also refactored the event a little bit. Now it consists of [event-name {message-attributes}]. That way you don't have to remember which argument is in which place, since now we can specify up to 4 things per message: title, description, flavor and ttl.