Using interceptor to inject current time into clean-old-messages event handler
As per Daniel's suggestion I've looked into interceptors. There was an example of injecting :now into an even handler, which is what I need for cleaning old messages. Before I was passing the current time as an event param, which is fine but with the interceptor I simly need to make sure I trigger the event without knowing what it needs to clean the messages.